Rework options menu.
It's much nicer now.
