+++ Content.Client/UserInterface/KeyRebindControl.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using Content.Client.UserInterface.Stylesheets;
+using Content.Shared.Input;
+using Robust.Client.Input;
+using Robust.Client.Interfaces.Input;
+using Robust.Client.UserInterface;
+using Robust.Client.UserInterface.Controls;
+using Robust.Shared.Input;
+using Robust.Shared.IoC;
+using Robust.Shared.Localization;
+using Robust.Shared.Timing;
+using Robust.Shared.Utility;
+
+#nullable enable
+
+namespace Content.Client.UserInterface
+{
+}
